Fixing display issues when separate/visible group mode is enabled.
When the grade_report_quick_edit class is created, it generates the group dropdown for the different groups in the default grouping.
But when it generates that dropdown it makes a call to the output renderer. The output renderer tries to figure out the theme and layout to use, but since the layout hasn't been defined yet, it uses the base layout. The base layout doesn't have block regions defined. So the page was crashing, because it is trying to put content in the side-pre region.
The fix can be to either add the side-pre region to the base theme layout or define the layout before the grade_report_quick_edit class is created. I am doing the later.